requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
